requirements.txt
requests
fastapi
gradio
python-dotenv
openai
uvicorn
gspread 
google-auth
sendgrid

